Stop retrying after a few errors
