+++ lib/mayu/client/src/Mayu.ts
@@ -4,8 +4,9 @@ import PingTimer from "./PingTimer.js";
 import type MayuPingElement from "./custom-elements/mayu-ping";
 import type MayuProgressBar from "./custom-elements/mayu-progress-bar";
 import defineCustomElements from "./custom-elements";
-import { info } from "console";
 defineCustomElements();
+
+import h from "./h";
 
 const SESSION_ID_KEY = "mayu.sessionId";
 
@@ -27,7 +28,10 @@ export default async function init(encryptedState: string) {
 
   const es = new EventSource(`/__mayu/session/${sessionId}/events`);
 
+  let errorCount = 0;
+
   es.onopen = () => {
+    errorCount = 0;
     console.log("Opened session", sessionId);
     sessionStorage.setItem(SESSION_ID_KEY, sessionId);
     startPing(es, sessionId);
@@ -40,8 +44,14 @@ export default async function init(encryptedState: string) {
   es.onerror = () => {
     console.log({ isUnloading, readyState: document.readyState });
     if (isUnloading) return;
+
     sessionStorage.removeItem(SESSION_ID_KEY);
     document.body.appendChild(disconnectedElement);
+
+    if (errorCount++ > 5) {
+      console.log("Closing event source");
+      es.close();
+    }
   };
 
   es.addEventListener("navigate", (e) => {
@@ -49,6 +59,25 @@ export default async function init(encryptedState: string) {
     console.log("Navigating to", path);
     history.pushState({}, "", path);
     //  progressBar.setAttribute("progress", "100");
+  });
+
+  es.addEventListener("exception", (e) => {
+    const error = JSON.parse(e.data) as {
+      type: string;
+      message: string;
+      backtrace: string[];
+    };
+    const { type, message, backtrace } = error;
+    const cleanedBacktrace = backtrace
+      .filter((line) => !/\/vendor\/bundle\//.test(line))
+      .join("\n");
+
+    const el = h("mayu-exception", [
+      h("span", [`${type}: ${message}`], { slot: "title" }),
+      h("span", [cleanedBacktrace], { slot: "backtrace" }),
+    ]);
+
+    document.body.appendChild(el);
   });
 
   window.addEventListener("popstate", () => {
@@ -175,15 +204,30 @@ async function startPing(es: EventSource, sessionId: string) {
   async function pingLoop() {
     while (true) {
       try {
-        const { ping, region } = await pingTimer.ping((now) => {
-          navigator.sendBeacon(
-            `/__mayu/session/${sessionId}/ping`,
-            JSON.stringify(now)
-          );
-        });
+        switch (es.readyState) {
+          case EventSource.OPEN: {
+            const { ping, region } = await pingTimer.ping((now) => {
+              navigator.sendBeacon(
+                `/__mayu/session/${sessionId}/ping`,
+                JSON.stringify(now)
+              );
+            });
 
-        pingElement.setAttribute("ping", `${ping} ms`);
-        pingElement.setAttribute("region", region);
+            pingElement.setAttribute("ping", `${ping} ms`);
+            pingElement.setAttribute("region", region);
+            break;
+          }
+          case EventSource.CLOSED: {
+            pingElement.setAttribute("ping", `N/A ms`);
+            pingElement.setAttribute("region", "closed");
+            break;
+          }
+          case EventSource.CONNECTING: {
+            pingElement.setAttribute("ping", `N/A ms`);
+            pingElement.setAttribute("region", "connecting");
+            break;
+          }
+        }
 
         await pingTimer.sleep(PingTimer.PING_FREQUENCY_MS);
       } catch (e) {
