+++ lib/mayu/client/src/custom-elements/mayu-exception.html
@@ -37,6 +37,7 @@
     flex: 0;
     background: transparent;
     font-size: 2em;
+    font-weight: bold;
     opacity: 0.5;
     border: 0;
     cursor: pointer;
@@ -47,7 +48,7 @@
   }
 
   h1 {
-    font-size: 2em;
+    font-size: 1.2em;
     margin: 0;
     padding: 0;
     flex: 1;
